Update CastDetailViewController.swift
collectionView's delegate, dataSource and DelegateFlowLayout protocols were added with the extension.
The contents of the cells were filled with the images array where image data is kept.
added gesture tab recognizer to biographtContainerView with setBiographyHeight func.
In the touchedBiography objc function, an algorithm was written to lengthen and shorten the BiographyContainerView based on the size of the BiographyLabel